refactor(cli): share provider-binding helpers between set and import
Extract `pickProviderRef` and a `writeSecret` helper into a shared
`secret-binding.ts` module so the set and import commands stop carrying
the same per-env lookup and the same `provider.set({ ... })` argument
construction.
